+++ backend/api/views.py
@@ -1,4 +1,4 @@
-from api.certificate_serializers import FavouriteSerializer
+from api.serializers.certificate_serializers import FavouriteSerializer
 from docs.models import Favourite
 from django.contrib.auth import get_user_model
 from djoser.views import UserViewSet as DjoserUserViewSet
